fix(aep-86): sync generated SDK outputs
Regenerate the TypeScript and Rust SDK bindings for the AEP-86 proto surface.
Keep the AEP-local lint fixes from the generated checks.

+++ go/sdl/verification_test.go
@@ -164,7 +164,7 @@ min_tier: 0
 	var v v2Verification
 	err := yaml.Unmarshal(in, &v)
 	require.NoError(t, err)
-	assert.Equal(t, 0, v.MinTier)
+	assert.Equal(t, int32(0), v.MinTier)
 	assert.Equal(t, "", v.AuditorMode)
 	assert.Empty(t, v.Capabilities)
 	assert.Empty(t, v.Auditors)
